Refactor auth package, add rate limiting, extract token parsing logic to be reusable in things that aren't gin
